+++ examples/caputo-gradient-spectral.py
@@ -32,26 +32,24 @@ def df(x: Array, alpha: float) -> Array:
     )
 
 
-from pycaputo.derivatives import CaputoDerivative, Side
+# instantiate the desired method
+from pycaputo.differentiation import caputo
 
-d = CaputoDerivative(order=0.9, side=Side.Left)
-
-from pycaputo import grad
-from pycaputo.differentiation import CaputoSpectralMethod
-
-method = CaputoSpectralMethod(d)
-
-from pycaputo.grid import make_jacobi_gauss_lobatto_points
+alpha = 0.9
+method = caputo.SpectralJacobi(alpha=0.9)
 
 # make a set of points at which to evaluate the gradient
 rng = np.random.default_rng(42)
 x = rng.uniform(3.0, 7.0, size=12)
 
 # evaluate the gradient
+from pycaputo import grad
+from pycaputo.grid import make_jacobi_gauss_lobatto_points
+
 p = make_jacobi_gauss_lobatto_points(32, a=0.0, b=1.0)
 df_num = grad(f, p, x, method=method)
 
-df_ref = df(x, d.order)
+df_ref = df(x, alpha)
 print(
     "Relative error: ",
     np.linalg.norm(df_num - df_ref) / np.linalg.norm(df_ref),