+++ src/pycaputo/differentiation/base.py
@@ -16,42 +16,24 @@ from pycaputo.utils import Array, ArrayOrScalarFunction
 class DerivativeMethod(ABC):
     """A generic method used to evaluate a fractional derivative at a set of points."""
 
-    d: FractionalOperator
-
-    if __debug__:
-
-        def __post_init__(self) -> None:
-            if not self.supports(self.d.order):
-                raise ValueError(
-                    f"Method '{self.name}' does not support derivative order "
-                    f"'{self.d.order}'"
-                )
-
     @property
     def name(self) -> str:
-        """An identifier for the differentiation method."""
+        """An (non-unique) identifier for the differentiation method."""
         return type(self).__name__.replace("Method", "")
 
     @property
     @abstractmethod
-    def order(self) -> float:
-        """Expected order of convergence of the method."""
-
-    @abstractmethod
-    def supports(self, alpha: float) -> bool:
-        """
-        :returns: *True* if the method supports computing the fractional
-            order derivative of order *alpha* and *False* otherwise.
-        """
+    def d(self) -> FractionalOperator:
+        """The fractional derivative that is being discretized by the method."""
 
 
 @singledispatch
 def diff(m: DerivativeMethod, f: ArrayOrScalarFunction, p: Points) -> Array:
-    """Evaluate the fractional derivative of *f* at *x*.
+    """Evaluate the fractional derivative of *f* at *p* using method *m*.
 
-    Note that not all numerical methods can evaluate the derivative at all
-    points in *x*. For example, the :class:`CaputoL1Method` cannot evaluate the
-    derivative at ``x[0]``.
+    This is a low-level function that should be implemented by methods. For
+    a higher-level function see :func:`pycaputo.diff`. Note that this function
+    evaluates the derivative at all points in *p*, not just at *p[-1]*.
 
     :arg m: method used to evaluate the derivative.
     :arg f: a simple function for which to evaluate the derivative.